task1/src/main/java/com/javaabuser/task/controllers/CategoryController.java
package com.javaabuser.task.controllers;

import com.javaabuser.task.dto.CategoryDTO;
import com.javaabuser.task.services.CategoryService;
import org.modelmapper.ModelMapper;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;
import java.util.stream.Collectors;

@RestController
@RequestMapping("/categories")
public class CategoryController {
    private final ModelMapper modelMapper;
    private final CategoryService categoryService;

    @Autowired
    public CategoryController(ModelMapper modelMapper, CategoryService categoryService) {
        this.modelMapper = modelMapper;
        this.categoryService = categoryService;
    }

    @GetMapping
    public List<CategoryDTO> getCategories(){
        return categoryService.getCategories()
                .stream()
                .map(category -> modelMapper.map(category, CategoryDTO.class))
                .collect(Collectors.toList());
    }
}
